refactor: separate tests for COSE structures
